Logout related changes and other changes like ExpenseType, Designation, validations on get update delete rest calls

+++ FS/src/main/java/com/fs/controller/base/BaseController.java
@@ -8,10 +8,12 @@ import javax.validation.Valid;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,6 +21,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import com.fs.constants.FSExceptionConstants;
+import com.fs.context.FSContext;
+import com.fs.pojo.Firm;
+import com.fs.pojo.base.FirmAware;
+import com.fs.pojo.base.Trackable;
 import com.fs.services.base.Service;
 import com.fs.utils.DataBinderUtil;
 import com.fs.utils.QueryParamUtil;
@@ -80,20 +87,71 @@ public abstract class BaseController<B, D> {
 	@RequestMapping(path = "/{id}", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Object> update(@Valid @RequestBody B object, @PathVariable Long id){
 		D d = getService().findById(id);
-		BeanUtils.copyProperties(object, d);
-		QueryParamUtil.setFirmData(d);
-		getService().save(d);
+		if (firmValidation(d)) {
+			throw new AccessDeniedException(FSExceptionConstants.DONT_HAVE_PERMISSION);
+		} else {
+			BeanUtils.copyProperties(object, d);
+			QueryParamUtil.setFirmData(d);
+			getService().save(d);
+		}
 		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 	
-	@SuppressWarnings({ "rawtypes", "unchecked" })
-	@RequestMapping(method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity delete(@RequestBody B object){
-		Object entity = binder.toDomain(object, getDomainObject());
-		D d = (D) entity;
-		getService().delete(d);
+	@SuppressWarnings({ "rawtypes"})
+	@RequestMapping(path = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity delete(@PathVariable Long id){
+		D d = getService().findById(id);
+		if (firmValidation(d)) {
+			throw new AccessDeniedException(FSExceptionConstants.DONT_HAVE_PERMISSION);
+		} else {
+			if(d instanceof Trackable) {
+				Trackable trackable = (Trackable) d;
+				trackable.setDeleteStatus(Boolean.TRUE);
+				getService().save(d);
+			} else {
+				getService().delete(d);
+			}
+		}
 		return new ResponseEntity(HttpStatus.OK);
 	}
 	
+	@SuppressWarnings({ "unchecked", "rawtypes" })
+	@RequestMapping(path = "/getGlobal", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+	public <T> ResponseEntity<List> getGlobal(@RequestParam Map<String, String> map) {
+		List<D> domainList = new ArrayList<>();
+		List<Object> list = new ArrayList<Object>();
+		
+		List<Example> examples = QueryParamUtil.getFirmExamples(getDomainObject());
+		for (Example example : examples) {
+			List<D> result = getService().findAll(example);
+			if (!result.isEmpty()) {
+				domainList.addAll(result);
+			}
+		}
+		
+		if (!domainList.isEmpty()) {
+			for (D d : domainList) {
+				list.add(binder.toBusiness(d, getBusinessObject()));
+			}
+		}
+
+		return new ResponseEntity<List>(list, HttpStatus.OK);
+	}
+	
+	private Boolean firmValidation(D d) {
+		if (d instanceof FirmAware) {
+			FirmAware obj = (FirmAware) d;
+			return !FSContext.getFirmId().equals(obj.getFirmId());
+		} else if(d instanceof Firm) {
+			Firm firm = (Firm) d;
+			if (FSContext.getFirmId().equals(1l)) {
+				return Boolean.FALSE;
+			} else {
+				return !FSContext.getFirmId().equals(firm.getId());
+			}
+		} else {
+			return Boolean.FALSE;
+		}
+	}
 	
 }
